migration(*): refactored sidebar logic to support unlimited menus
