+++ package.json
@@ -18,17 +18,18 @@
     "@tinymce/tinymce-vue": "^2.0.0",
     "@vue/composition-api": "^0.5.0",
     "axios": "~0.18.0",
-    "element-plus": "^1.0.1-alpha.14",
+    "core-js": "^3.6.5",
+    "element-plus": "^1.0.1-alpha.21",
     "event-source-polyfill": "^1.0.7",
     "fastscan": "^1.0.4",
     "good-storage": "^1.1.0",
     "js-cookie": "^2.2.0",
     "lodash": "^4.17.14",
+    "mitt": "^2.1.0",
     "moment": "^2.24.0",
     "photoswipe": "^4.1.2",
     "screenfull": "^4.2.0",
     "tinymce": "^5.0.1",
-    "core-js": "^3.6.5",
     "vue": "^3.0.0",
     "vue-awesome-swiper": "^3.1.3",
     "vue-croppa": "^1.3.8",
