+++ src/component/layout/index.js
@@ -1,9 +1,9 @@
 import NavBar from './nav-bar'
-import SideBar from './side-bar'
+import Sidebar from './sidebar/sidebar.vue'
 import AppMain from './app-main'
 import ReuseTab from './reuse-tab'
 import MenuTab from './menu-tab.vue'
 import BackTop from './back-top.vue'
 import User from './user.vue'
 
-export { NavBar, SideBar, AppMain, ReuseTab, MenuTab, BackTop, User }
+export { NavBar, Sidebar, AppMain, ReuseTab, MenuTab, BackTop, User }
