+++ src/component/layout/sidebar/logo.vue
@@ -1,0 +1,55 @@
+<template>
+  <div class="logo" v-if="!elMenuCollapse"><img src="../../../assets/image/logo.png" alt="" /></div>
+  <div class="mobile-logo" v-else><img src="../../../assets/image/mobile-logo.png" alt="" /></div>
+</template>
+
+<script>
+export default {
+  props: {
+    elMenuCollapse: {
+      type: Boolean,
+      required: true,
+    },
+  },
+}
+</script>
+
+<style lang="scss" scoped>
+.logo {
+  width: $sidebar-width;
+  height: $header-height;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 30px;
+  color: #fff;
+  transition: all 0.5s ease-in-out;
+  background-color: #122150;
+  transition: all 0.3s linear;
+  position: sticky;
+  top: 0;
+  left: 0;
+  z-index: 99;
+
+  img {
+    width: 120px;
+    transition: all 0.3s linear;
+  }
+}
+
+.mobile-logo {
+  width: 64px;
+  height: 86px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #122150;
+  transition: all 0.3s linear;
+
+  img {
+    width: 40px;
+    height: 40px;
+    transition: all 0.3s linear;
+  }
+}
+</style>
