+++ src/lin/util/emitter.js
@@ -1,0 +1,3 @@
+import mitt from 'mitt'
+
+export default mitt()
