+++ src/plugin/custom/stage-config.js
@@ -15,7 +15,7 @@ const CustomRouter = {
       route: '/custom/upload-image',
       filePath: 'plugin/custom/view/upload-image.vue',
       inNav: true,
-      icon: 'iconfont icon-zidingyi',
+      icon: 'iconfont icon-upload',
       permission: null,
     },
     {
@@ -25,17 +25,17 @@ const CustomRouter = {
       route: '/custom/gallery',
       filePath: 'plugin/custom/view/gallery.vue',
       inNav: true,
-      icon: 'iconfont icon-zidingyi',
+      icon: 'iconfont icon-art-gallery-',
       permission: null,
     },
     {
-      title: '富文本',
+      title: 'editor 富文本',
       type: 'view',
       name: 'Tinymce',
       route: '/custom/tinymce',
       filePath: 'plugin/custom/view/tinymce.vue',
       inNav: true,
-      icon: 'iconfont icon-zidingyi',
+      icon: 'iconfont icon-fuwenbenbianjiqi_gongshi',
       permission: null,
     },
     {
@@ -45,7 +45,7 @@ const CustomRouter = {
       route: '/custom/multiple',
       filePath: 'plugin/custom/view/multiple-input.vue',
       inNav: true,
-      icon: 'iconfont icon-zidingyi',
+      icon: 'iconfont icon-multiple_inputs',
       permission: null,
     },
   ],
