+++ src/view/about/about.vue
@@ -173,8 +173,6 @@ import { ref, onMounted } from 'vue'
 
 export default {
   setup() {
-    console.log(1111)
-
     const showTeam = ref(false)
     const activeName = ref('first')
     const { clientWidth } = document.body
